+++ lib/keycloak-admin/client/configurable_token_client.rb
@@ -18,7 +18,7 @@ module KeycloakAdmin
     def exchange_with(user_access_token, token_lifespan_in_seconds)
       response = execute_http do
         Resource.execute(
-          @configuration.faraday_options.merge(
+          connection_options.merge(
             method: :post,
             url: token_url,
             payload: { tokenLifespanInSeconds: token_lifespan_in_seconds }.to_json,